Start breaking up code into independent modules
