+++ src/conn/mod.rs
@@ -1,0 +1,3 @@
+pub mod common;
+pub mod usb;
+pub mod mock;
