+++ src/network/mod.rs
@@ -1,0 +1,3 @@
+pub mod common;
+pub mod server;
+pub mod client;
